doc: update api doc.

+++ app/src/main/java/com/cmoney/kolfanci/model/usecase/ChatRoomUseCase.kt
@@ -65,7 +65,7 @@ class ChatRoomUseCase(
         val votingIds = attachment.filter {
             it.first == AttachmentType.Choice && it.second.other is VoteModel
         }.map {
-            (it.second.other as VoteModel).id.toLong()
+            (it.second.other as VoteModel).id
         }
 
         val chatMessageParam = ChatMessageParam(
